split VepAnnotationFileWriter (VepProcess)
