+++ src/test/java/uk/ac/ebi/eva/pipeline/io/writers/VepAnnotationFileWriterTest.java
@@ -45,6 +45,10 @@ public class VepAnnotationFileWriterTest {
 
     private static final long TIMEOUT_IN_SECONDS = 5L;
 
+    /**
+     * the mockvep writes an extra line as if some variant had two annotations, to check that the writer is not assuming
+     * that the count of variants to annotate is the same as variantAnnotations to write in the file.
+     */
     private static final int EXTRA_ANNOTATIONS = 1;
 
     private AnnotationParameters annotationParameters;
